jobs/data-processor/requirements.txt
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
requests==2.31.0
cryptography==41.0.7
python-dotenv==1.0.0
alembic==1.13.1 